requirements.txt
anyio==3.7.0
asyncio==3.4.3
click==8.1.3
fastapi==0.98.0
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pydantic==1.10.9
PyMySQL==1.1.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
strawberry-graphql==0.189.3
typing_extensions==4.6.3
uvicorn==0.22.0
